+++ sample/src/main/java/net/xpece/android/support/preference/sample/SampleApplication.java
@@ -1,0 +1,26 @@
+package net.xpece.android.support.preference.sample;
+
+import android.app.Application;
+import android.os.StrictMode;
+
+import com.squareup.leakcanary.LeakCanary;
+
+/**
+ * Created by Eugen on 15.02.2017.
+ */
+
+public class SampleApplication extends Application {
+    @Override
+    public void onCreate() {
+        super.onCreate();
+
+        StrictMode.setVmPolicy(new StrictMode.VmPolicy.Builder()
+            .detectAll()
+            .penaltyLog()
+            .build());
+
+        if (!LeakCanary.isInAnalyzerProcess(this)) {
+            LeakCanary.install(this);
+        }
+    }
+}